requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
apscheduler==3.6.3
asgiref==3.4.1; python_version >= '3.6'
cachetools==4.2.2; python_version ~= '3.5'
certifi==2021.5.30
charset-normalizer==2.0.4; python_version >= '3'
django==3.2.6
idna==3.2; python_version >= '3'
python-telegram-bot==13.7
pytz==2021.1
requests==2.26.0
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.4.1; python_version >= '3.5'
tornado==6.1; python_version >= '3.5'
tzlocal==2.1
urllib3==1.26.6; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
